Moved SearchTest under to Collection
