+++ tests/Cms/Collections/SearchTest.php
@@ -103,4 +103,66 @@ class SearchTest extends TestCase
         $search = Search::collection($collection, 'G', ['minlength' => 1]);
         $this->assertCount(1, $search);
     }
+
+    public function app(): App
+    {
+        return new App([
+            'roots' => [
+                'index' => '/dev/null'
+            ],
+            'site' => [
+                'children' => [
+                    [
+                        'slug'  => 'categories',
+                        'files' => [
+                            ['filename' => 'phone.jpg'],
+                            ['filename' => 'cell-phone.jpg'],
+                            ['filename' => 'computer.jpg']
+                        ]
+                    ],
+                    [
+                        'slug'  => 'products',
+                        'files' => [
+                            ['filename' => 'apple.jpg'],
+                            ['filename' => 'samsung.jpg']
+                        ]
+                    ],
+                    [
+                        'slug' => 'contact'
+                    ]
+                ],
+                'files' => [
+                    ['filename' => 'website.jpg']
+                ]
+            ],
+            'users' => [
+                ['email' => 'admin@getkirby.com'],
+                ['email' => 'editor@getkirby.com'],
+                ['email' => 'user1@getkirby.com'],
+                ['email' => 'user2@getkirby.com'],
+                ['email' => 'user3@getkirby.com']
+            ],
+        ]);
+    }
+
+    public function testFiles()
+    {
+        $this->assertSame(5, $this->app()->site()->index()->files()->count());
+        $this->assertInstanceOf(Files::class, $files = Search::files('phone'));
+        $this->assertSame(2, $files->count());
+    }
+
+    public function testPages()
+    {
+        $this->assertSame(3, $this->app()->site()->index()->count());
+        $this->assertInstanceOf(Pages::class, $pages = Search::pages('products'));
+        $this->assertSame(1, $pages->count());
+    }
+
+    public function testUsers()
+    {
+        $this->assertSame(5, $this->app()->users()->count());
+        $this->assertInstanceOf(Users::class, $users = Search::users('user'));
+        $this->assertSame(3, $users->count());
+    }
 }
